# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Nexus REST API",
    "description": "Explore resources and get their data.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/v1/jobs/export": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Creates a new export job.",
        "operationId": "Jobs_Export",
        "requestBody": {
          "x-name": "parameters",
          "description": "Export parameters.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportParameters"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/clear-cache": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Clears the catalog cache for the specified period of time.",
        "operationId": "Jobs_ClearCache",
        "parameters": [
          {
            "name": "catalogId",
            "in": "query",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "begin",
            "in": "query",
            "required": true,
            "description": "Start date/time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 2
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "description": "End date/time.",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{jobId}/status": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Gets the status of the specified job.",
        "operationId": "Jobs_GetJobStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}": {
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancels the specified job.",
        "operationId": "Jobs_DeleteJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Job": {
        "type": "object",
        "description": "Description of a job.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "06f8eb30-5924-4a71-bdff-322f92343f5b",
            "format": "guid"
          },
          "type": {
            "type": "string",
            "description": "export"
          },
          "owner": {
            "type": "string",
            "description": "test@nexus.localhost"
          },
          "parameters": {
            "description": "Job parameters.",
            "nullable": true
          }
        }
      },
      "ExportParameters": {
        "type": "object",
        "description": "A structure for export parameters.",
        "additionalProperties": false,
        "properties": {
          "begin": {
            "type": "string",
            "description": "2020-02-01T00:00:00Z",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "description": "2020-02-02T00:00:00Z",
            "format": "date-time"
          },
          "filePeriod": {
            "type": "string",
            "description": "00:00:00",
            "format": "duration"
          },
          "type": {
            "type": "string",
            "description": "Nexus.Writers.Csv"
          },
          "resourcePaths": {
            "type": "array",
            "description": "[\"/IN_MEMORY/TEST/ACCESSIBLE/T1/1_s_mean\", \"/IN_MEMORY/TEST/ACCESSIBLE/V1/1_s_mean\"]",
            "items": {
              "type": "string"
            }
          },
          "configuration": {
            "type": "object",
            "description": "{ \"RowIndexFormat\": \"Index\", \"SignificantFigures\": \"4\" }",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "JobStatus": {
        "type": "object",
        "description": "Describes the status of the job.",
        "additionalProperties": false,
        "properties": {
          "start": {
            "type": "string",
            "description": "The start date/time.",
            "format": "date-time"
          },
          "status": {
            "description": "The status.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/TaskStatus"
              }
            ]
          },
          "progress": {
            "type": "number",
            "description": "The progress from 0 to 1.",
            "format": "double"
          },
          "exceptionMessage": {
            "type": "string",
            "description": "An optional exception message.",
            "nullable": true
          },
          "result": {
            "description": "The optional result.",
            "nullable": true
          }
        }
      },
      "TaskStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Created",
          "WaitingForActivation",
          "WaitingToRun",
          "Running",
          "WaitingForChildrenToComplete",
          "RanToCompletion",
          "Canceled",
          "Faulted"
        ],
        "enum": [
          "Created",
          "WaitingForActivation",
          "WaitingToRun",
          "Running",
          "WaitingForChildrenToComplete",
          "RanToCompletion",
          "Canceled",
          "Faulted"
        ]
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Please enter 'Bearer {your JWT token}'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}